requirements.txt
dj-database-url==0.4.2
Django==2.0.1
djangorestframework==3.7.7
Faker==0.8.8
freezegun==0.3.9
mixer==6.0.1
prettyconf==1.2.3
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
text-unidecode==1.1
